Attached health component to the dodgeball character

+++ Source/DodgeBall/HealthComponent.h
@@ -16,14 +16,14 @@ public:
 	// Sets default values for this component's properties
 	UHealthComponent();
 	
-	void LoseHealth(float Damage);
+	void LoseHealth(int Damage);
 	
 protected:
 	// Called when the game starts
 	virtual void BeginPlay() override;
 	
 	UPROPERTY(EditDefaultsOnly, Category = Health)
-	float Health = 100.f;
+	int Health = 100;
 public:	
 	// Called every frame
 	virtual void TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction) override;
